refactor(components): split into separate files & rename action | hj | #3537
