+++ packages/components/src/utils/select/index.ts
@@ -1,3 +1,5 @@
-export * from './select-dropdown';
-export * from './keyboard-behavior';
 export * from './filter';
+export * from './getMultiSelectActionFromKeyboardEvent';
+export * from './getSelectActionFromKeyboardEvent';
+export * from './keyboard-behavior';
+export * from './select-dropdown';
